fix(cli): fix status of prompt not updating
if the client id that was used to submit the prompt is not the one
from cmfy, we do not recieve progress or execution messages. We
simply recieve status update with the queue. So the entries were
never updated.

refactor: trying to find a good factoring for this... not easy